refactor: 개별 클래스로 파일 분리
