+++ android/app/src/main/java/app/priceguard/data/dto/detail/ProductDetailResult.kt
@@ -1,0 +1,16 @@
+package app.priceguard.data.dto.detail
+
+import app.priceguard.data.graph.ProductChartData
+
+data class ProductDetailResult(
+    val productName: String,
+    val productCode: String,
+    val shop: String,
+    val imageUrl: String,
+    val rank: Int,
+    val shopUrl: String,
+    val targetPrice: Int,
+    val lowestPrice: Int,
+    val price: Int,
+    val priceData: List<ProductChartData>
+)
