+++ android/app/src/main/java/app/priceguard/data/dto/detail/ProductResponse.kt
@@ -1,7 +1,6 @@
 package app.priceguard.data.dto.detail
 
 import app.priceguard.data.dto.PriceDataDTO
-import app.priceguard.data.graph.ProductChartData
 import kotlinx.serialization.Serializable
 
 @Serializable
@@ -18,17 +17,4 @@ data class ProductResponse(
     val lowestPrice: Int? = null,
     val price: Int? = null,
     val priceData: List<PriceDataDTO>? = null
-)
-
-data class ProductDetailResult(
-    val productName: String,
-    val productCode: String,
-    val shop: String,
-    val imageUrl: String,
-    val rank: Int,
-    val shopUrl: String,
-    val targetPrice: Int,
-    val lowestPrice: Int,
-    val price: Int,
-    val priceData: List<ProductChartData>
 )
