+++ android/app/src/main/java/app/priceguard/data/dto/list/ProductDTO.kt
@@ -1,7 +1,6 @@
 package app.priceguard.data.dto.list
 
 import app.priceguard.data.dto.PriceDataDTO
-import app.priceguard.data.graph.ProductChartData
 import kotlinx.serialization.Serializable
 
 @Serializable
@@ -13,14 +12,4 @@ data class ProductDTO(
     val targetPrice: Int? = null,
     val price: Int? = null,
     val priceData: List<PriceDataDTO>? = null
-)
-
-data class ProductData(
-    val productName: String,
-    val productCode: String,
-    val shop: String,
-    val imageUrl: String,
-    val targetPrice: Int,
-    val price: Int,
-    val priceData: List<ProductChartData>
 )
