+++ android/app/src/main/java/app/priceguard/data/dto/list/ProductData.kt
@@ -1,0 +1,13 @@
+package app.priceguard.data.dto.list
+
+import app.priceguard.data.graph.ProductChartData
+
+data class ProductData(
+    val productName: String,
+    val productCode: String,
+    val shop: String,
+    val imageUrl: String,
+    val targetPrice: Int,
+    val price: Int,
+    val priceData: List<ProductChartData>
+)
