+++ android/app/src/main/java/app/priceguard/data/dto/login/LoginResult.kt
@@ -1,0 +1,7 @@
+package app.priceguard.data.dto.login
+
+data class LoginResult(
+    val loginState: LoginState,
+    val accessToken: String?,
+    val refreshToken: String?
+)
