+++ android/app/src/main/java/app/priceguard/data/dto/login/LoginState.kt
@@ -1,0 +1,7 @@
+package app.priceguard.data.dto.login
+
+enum class LoginState {
+    SUCCESS,
+    INVALID_PARAMETER,
+    UNDEFINED_ERROR
+}
