+++ android/app/src/main/java/app/priceguard/data/dto/recommend/RecommendProductDTO.kt
@@ -1,7 +1,6 @@
 package app.priceguard.data.dto.recommend
 
 import app.priceguard.data.dto.PriceDataDTO
-import app.priceguard.data.graph.ProductChartData
 import kotlinx.serialization.Serializable
 
 @Serializable
@@ -13,14 +12,4 @@ data class RecommendProductDTO(
     val price: Int? = null,
     val rank: Int? = null,
     val priceData: List<PriceDataDTO>? = null
-)
-
-data class RecommendProductData(
-    val productName: String,
-    val productCode: String,
-    val shop: String,
-    val imageUrl: String,
-    val price: Int,
-    val rank: Int,
-    val priceData: List<ProductChartData>
 )
