+++ android/app/src/main/java/app/priceguard/data/dto/renew/RenewResult.kt
@@ -1,0 +1,8 @@
+package app.priceguard.data.dto.renew
+
+enum class RenewResult {
+    SUCCESS,
+    EXPIRED,
+    UNAUTHORIZED,
+    UNKNOWN_ERROR
+}
