+++ android/app/src/main/java/app/priceguard/data/dto/signup/SignupResult.kt
@@ -1,0 +1,7 @@
+package app.priceguard.data.dto.signup
+
+data class SignupResult(
+    val signUpState: SignupState,
+    val accessToken: String?,
+    val refreshToken: String?
+)
