+++ android/app/src/main/java/app/priceguard/data/dto/signup/SignupState.kt
@@ -1,0 +1,8 @@
+package app.priceguard.data.dto.signup
+
+enum class SignupState {
+    SUCCESS,
+    INVALID_PARAMETER,
+    DUPLICATE_EMAIL,
+    UNDEFINED_ERROR
+}
