+++ android/app/src/main/java/app/priceguard/data/dto/signup/UserDataDTO.kt
@@ -10,8 +10,3 @@ data class UserDataDTO(
     val iat: Int,
     val exp: Int
 )
-
-data class UserDataResult(
-    val email: String,
-    val name: String
-)
