+++ android/app/src/main/java/app/priceguard/data/dto/signup/UserDataResult.kt
@@ -1,0 +1,6 @@
+package app.priceguard.data.dto.signup
+
+data class UserDataResult(
+    val email: String,
+    val name: String
+)
